requirements.txt
numpy
matplotlib
pillow
tensorflow
streamlit